Initial post
This is the initial post of the Require.JS port. Please note that this
is not free-ware. We would love your feedback and improvements, but this
code is a part of our RTL. Also note that the require.js file belongs to
the original js developer (see file for details)